reorganize things, add cli documentation
