+++ fluvii/__init__.py
@@ -1,4 +1,14 @@
-from .fluvii_app import FluviiAppFactory, FluviiTableAppFactory, FluviiMultiMessageAppFactory, FluviiAppConfig
+from .components import (
+    AuthKafkaConfig,
+    ConsumerFactory,
+    ConsumerConfig,
+    ProducerFactory,
+    ProducerConfig,
+    SchemaRegistry,
+    SchemaRegistryConfig
+)
+from .apps import FluviiAppFactory, FluviiTableAppFactory, FluviiMultiMessageAppFactory, FluviiAppConfig
+from .kafka_tools.fluvii_toolbox import FluviiToolbox
 from .exceptions import WrappedSignals
 from .logging_utils import init_logger
