+++ fluvii/components/producer/__init__.py
@@ -1,0 +1,3 @@
+from .producer import Producer
+from .config import ProducerConfig
+from .producer_factory import ProducerFactory
