+++ fluvii/components/producer/config.py
@@ -1,6 +1,6 @@
 from typing import Optional
 from fluvii.config_bases import KafkaConfigBase, FluviiConfigBase
-from fluvii.auth import AuthKafkaConfig, get_auth_kafka_config
+from fluvii.components.auth import AuthKafkaConfig, get_auth_kafka_config
 
 
 class ProducerConfig(KafkaConfigBase, FluviiConfigBase):
