--- fluvii/consumer/__init__.py
@@ -1,2 +1,0 @@
-from .consumer import Consumer, TransactionalConsumer
-from .config import ConsumerConfig
