+++ fluvii/logging_utils.py
@@ -19,7 +19,7 @@ def _init_handler(loglevel=None):
 
 
 def _get_loglevel():
-    return environ.get('FLUVII_LOGLEVEL', 'DEBUG')
+    return environ.get('FLUVII_LOGLEVEL', 'INFO')
 
 
 def init_logger(name, loglevel=None):
