--- fluvii/producer/__init__.py
@@ -1,2 +1,0 @@
-from .producer import Producer, TransactionalProducer
-from .config import ProducerConfig
