--- fluvii/transaction/__init__.py
@@ -1,1 +1,0 @@
-from .transaction import Transaction, TableTransaction
